move Plate into correct folder
